treewide: Refactor AppConstant references for 41affb1
