+++ src/test/java/com/unconv/spring/common/ExceptionHandling.java
@@ -1,6 +1,6 @@
 package com.unconv.spring.common;
 
-import static com.unconv.spring.utils.AppConstants.PROFILE_TEST;
+import static com.unconv.spring.consts.AppConstants.PROFILE_TEST;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.web.bind.annotation.ControllerAdvice;
