+++ src/test/java/com/unconv/spring/web/controllers/SensorSystemControllerTest.java
@@ -1,9 +1,9 @@
 package com.unconv.spring.web.controllers;
 
+import static com.unconv.spring.consts.AppConstants.DEFAULT_SS_SORT_BY;
+import static com.unconv.spring.consts.AppConstants.DEFAULT_SS_SORT_DIRECTION;
+import static com.unconv.spring.consts.AppConstants.PROFILE_TEST;
 import static com.unconv.spring.consts.MessageConstants.ENVT_RECORD_ACCEPTED;
-import static com.unconv.spring.utils.AppConstants.DEFAULT_SS_SORT_BY;
-import static com.unconv.spring.utils.AppConstants.DEFAULT_SS_SORT_DIRECTION;
-import static com.unconv.spring.utils.AppConstants.PROFILE_TEST;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.Matchers.hasSize;
