+++ src/test/java/com/unconv/spring/web/controllers/UnconvUserControllerTest.java
@@ -1,8 +1,8 @@
 package com.unconv.spring.web.controllers;
 
+import static com.unconv.spring.consts.AppConstants.DEFAULT_PAGE_SIZE;
+import static com.unconv.spring.consts.AppConstants.PROFILE_TEST;
 import static com.unconv.spring.consts.MessageConstants.USER_CREATE_SUCCESS;
-import static com.unconv.spring.utils.AppConstants.DEFAULT_PAGE_SIZE;
-import static com.unconv.spring.utils.AppConstants.PROFILE_TEST;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.Matchers.hasSize;
